+++ packages/client/src/composables/onContentUpdated.ts
